improve handling of auth removal

+++ components/fetch/fetch-self.ts
@@ -14,26 +14,35 @@ export const fetchSelf = async (authToken: string) => {
       },
     },
   );
-  const person = (await personResponse.json()) as gapi.client.people.Person;
   if (!personResponse.ok) {
     RollbarErrorTracking.logErrorInfo(JSON.stringify(params));
     RollbarErrorTracking.logErrorInRollbar(personResponse.statusText);
+
+    if (personResponse.status === 401) {
+      // Reload if the current user fetch fails
+      window.location.reload();
+    }
   }
+
+  const person = (await personResponse.json()) as gapi.client.people.Person;
 
   // Handle multiple google accounts.
   // We store the google id used when fetching the current user and then compare
   // that to what we get in subsequent fetches.
-  const lastUpdatedUserId = localStorage.getItem('kelpLastUpdatedUserId');
-  if (!lastUpdatedUserId && person.resourceName) {
-    localStorage.setItem('kelpLastUpdatedUserId', person.resourceName);
-  } else if (person.resourceName && lastUpdatedUserId !== person.resourceName) {
-    try {
-      await deleteDatabase('production');
-    } catch (e) {
-      alert(JSON.stringify(e));
+  // NOTE: This doesn't appear to work very well. Perhaps things that are being saved while deleting the db?
+  if (typeof localStorage === 'object') {
+    const lastUpdatedUserId = localStorage.getItem('kelpLastUpdatedUserId');
+    if (!lastUpdatedUserId && person.resourceName) {
+      localStorage.setItem('kelpLastUpdatedUserId', person.resourceName);
+    } else if (person.resourceName && lastUpdatedUserId !== person.resourceName) {
+      try {
+        await deleteDatabase('production');
+      } catch (e) {
+        alert(JSON.stringify(e));
+      }
+      localStorage.setItem('kelpLastUpdatedUserId', person.resourceName);
+      window.location.reload();
     }
-    localStorage.setItem('kelpLastUpdatedUserId', person.resourceName);
-    window.location.reload();
   }
 
   return formatGooglePeopleResponse(person, person.resourceName);
